// src/app/generated/model/storageLocationRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StorageLocationRequest { 
    /**
     * Identifier of the warehouse this location belongs to
     */
    warehouseId: number;
    /**
     * Code identifying the location within its warehouse, e.g. a rack or box label
     */
    code: string;
}