// src/app/generated/model/warehouseUpdateRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WarehouseUpdateRequest { 
    /**
     * Display name of the warehouse
     */
    name?: string;
    /**
     * Whether the warehouse is currently active
     */
    active?: boolean;
}